+++ geotrek/outdoor/parsers.py
@@ -108,12 +108,14 @@ class GeotrekSiteParser(GeotrekOutdoorParser):
         "type": "outdoor_sitetype",
     }
     url_categories = {
+        "structure": "structure",
         "themes": "theme",
         'labels': 'label',
         'source': 'source',
         'managers': 'organism',
     }
     categories_keys_api_v2 = {
+        "structure": "name",
         "practice": "name",
         "sector": "name",
         "rating": "name",
@@ -126,6 +128,7 @@ class GeotrekSiteParser(GeotrekOutdoorParser):
         'managers': 'name',
     }
     natural_keys = {
+        "structure": "name",
         "practice": "name",
         "themes": "label",
         "type": "name",
@@ -191,12 +194,14 @@ class GeotrekCourseParser(GeotrekOutdoorParser):
         "type": "outdoor_coursetype",
     }
     url_categories = {
+        "structure": "structure",
         "themes": "theme",
         'labels': 'label',
         'source': 'source',
         'managers': 'organism',
     }
     categories_keys_api_v2 = {
+        "structure": "name",
         "themes": "label",
         "type": "name",
         'labels': 'name',
@@ -204,6 +209,7 @@ class GeotrekCourseParser(GeotrekOutdoorParser):
         'managers': 'name',
     }
     natural_keys = {
+        "structure": "name",
         "themes": "label",
         "type": "name",
         'labels': 'name',