+++ docs/changelog.rst
@@ -22,6 +22,7 @@ CHANGELOG
 - Remove unused `postgres` volume in docker compose file for production
 - Add `report` model on Intervention filters in module list (#3972)
 - Homogenize `structure` field in APIv2 to always return ids instead of names (#3007)
+- Add structures to Aggregator (#3569)
 
 
 2.104.2 (2024-04-04)